Split TextSearchBar into new file
